projekt2/pyproject.toml
[tool.poetry]
name = "projekt2"
version = "0.1.0"
description = ""
authors = ["Jan Kośnic"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.9.4"
pillow = "^11.0.0"
ascii-graph = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
